Extract superclass for DndResources to extend CustomSongResources
